Organize pages for apartments
